Extract shared YAML-edit primitives into _yaml_edit module
Six rule modules (CL-0003/0005/0007/0009/0014/0015) imported fixer
primitives — normalize_security_opt, DISABLED_SECURITY_PROFILES,
is_anchored_or_merged, line_indent, block_span, delete_lines, and the
block-shaping helpers — directly from fix.py, reaching into the fix
engine's internals to implement their fix() methods.

Promote the shared primitives into a dedicated compose_lint._yaml_edit
module. The six rules now import them from there rather than from fix; fix
consumes the same module, so it is the single home of the "shared helper
layer so individual rules stay small" (ADR-014, Part 1). fix.py drops from
860 to 576 lines.

Module placement note: the issue suggested rules/_yaml_edit_helpers.py, but
fix.py uses these primitives too, and a module under rules/ would make
fix.py import from the rules subpackage (a layering inversion). A neutral
top-level module both fix and the rules import cleanly avoids that; trivially
relocatable to rules/ if preferred.

Pure refactor: ruff/mypy clean, full suite (1133) green. test_fix.py's
primitive imports are split to the new module.

Closes #377
