+++ src/compose_lint/rules/CL0003_no_new_privileges.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
 
-from compose_lint.fix import (
+from compose_lint._yaml_edit import (
     DISABLED_SECURITY_PROFILES,
     block_span,
     extends_targets,
